+++ engine/samples/render/main.cpp
@@ -1,24 +1,9 @@
 #include <cubos/core/reflection/external/primitives.hpp>
 
 #include <cubos/engine/assets/plugin.hpp>
-#include <cubos/engine/render/bloom/plugin.hpp>
-#include <cubos/engine/render/camera/plugin.hpp>
-#include <cubos/engine/render/deferred_shading/plugin.hpp>
-#include <cubos/engine/render/depth/plugin.hpp>
-#include <cubos/engine/render/g_buffer/plugin.hpp>
-#include <cubos/engine/render/g_buffer_rasterizer/plugin.hpp>
-#include <cubos/engine/render/hdr/plugin.hpp>
+#include <cubos/engine/render/defaults/plugin.hpp>
 #include <cubos/engine/render/lights/environment.hpp>
-#include <cubos/engine/render/lights/plugin.hpp>
-#include <cubos/engine/render/mesh/plugin.hpp>
-#include <cubos/engine/render/picker/plugin.hpp>
-#include <cubos/engine/render/shader/plugin.hpp>
-#include <cubos/engine/render/split_screen/plugin.hpp>
-#include <cubos/engine/render/ssao/plugin.hpp>
-#include <cubos/engine/render/target/plugin.hpp>
-#include <cubos/engine/render/tone_mapping/plugin.hpp>
 #include <cubos/engine/render/voxels/palette.hpp>
-#include <cubos/engine/render/voxels/plugin.hpp>
 #include <cubos/engine/scene/plugin.hpp>
 #include <cubos/engine/settings/plugin.hpp>
 #include <cubos/engine/transform/plugin.hpp>
@@ -40,23 +25,7 @@ int main()
     cubos.plugin(transformPlugin);
     cubos.plugin(assetsPlugin);
     cubos.plugin(voxelsPlugin);
-    cubos.plugin(shaderPlugin);
-    cubos.plugin(renderTargetPlugin);
-    cubos.plugin(gBufferPlugin);
-    cubos.plugin(renderPickerPlugin);
-    cubos.plugin(renderDepthPlugin);
-    cubos.plugin(hdrPlugin);
-    cubos.plugin(cameraPlugin);
-    cubos.plugin(renderVoxelsPlugin);
-    cubos.plugin(renderMeshPlugin);
-    cubos.plugin(gBufferRasterizerPlugin);
-    cubos.plugin(lightsPlugin);
-    cubos.plugin(ssaoPlugin);
-    cubos.plugin(deferredShadingPlugin);
-    cubos.plugin(toneMappingPlugin);
-    cubos.plugin(splitScreenPlugin);
-    cubos.plugin(bloomPlugin);
-    cubos.tag(bloomTag).after(deferredShadingTag);
+    cubos.plugin(renderDefaultsPlugin);
     /// [Adding the plugins]
 
     cubos.plugin(scenePlugin);