+++ CHANGELOG.md
@@ -54,6 +54,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Split screen for the new renderer (#1149, **@tomas7770**).
 - Trait for representing inheritance relationships under reflection. (#693, **@roby2014**).
 - Bloom plugin (#1089, **@RiscadoA**).
+- Render Defaults plugin (#1158, **@RiscadoA**).
 
 ### Changed